requirements.txt
biopython
pytorch
numpy
pandas
tqdm